Move more miscellaneous functions into utils. Move metasprite constants into constants.h
